kill -9 으로 바꿔보자

+++ src/main/java/com/minsoo/co/tireerpserver/entity/purchase/PurchaseContent.java
@@ -36,10 +36,10 @@ public class PurchaseContent extends BaseEntity {
     private Stock stock;
 
     @Column(name = "price", nullable = false)
-    private Long price;
+    private Integer price;
 
     @Column(name = "quantity", nullable = false)
-    private Integer quantity;
+    private Long quantity;
 
     private PurchaseContent(Purchase purchase, TireDot tireDot, Stock stock) {
         this.purchase = purchase;
